move hyperlink code to different file
