+++ src/handlers/hyperlink.ts
@@ -1,163 +1,296 @@
-export class HyperlinkHandler {
-    private editor: HTMLDivElement;
-    private modal: HTMLDivElement | null = null;
-    private overlay: HTMLDivElement | null = null;
-    private currentAnchor: HTMLAnchorElement | null = null; // For update functionality
-    private savedRange: Range | null = null; // To save the selection range
-  
-    constructor(editor: HTMLDivElement) {
-      this.editor = editor;
-  
-      // Capture the caret/selection when the user clicks in the editor
-      this.editor.addEventListener('mouseup', () => this.saveSelection());
-      this.editor.addEventListener('keyup', () => this.saveSelection());
+import { blockType } from "../types/pieces";
+import { saveSelection,restoreSelection,getSelectionRange } from "../utils/selectionManager";
+import EditorView from "../view/editorView";
+import TextDocument from "../textDocument";
+class HyperlinkHandler {
+  savedSelection: { start: number; end: number } | null = null;
+  editorContainer: HTMLElement | null;
+  editorView: EditorView;
+  document:TextDocument;
+
+  constructor(editorContainer: HTMLElement, editorView: EditorView,document:TextDocument) {
+    this.editorContainer = editorContainer;
+    this.editorView = editorView;
+    this.document = document;
+  }
+
+  // Method to save the current selection or caret position
+
+  hanldeHyperlinkClick(
+    start: number,
+    end: number,
+    currentOffset: number,
+    selectedBlockId: string | null,
+    blocks: blockType
+  ) {
+    const existingLink = this.getCommonHyperlinkInRange(
+      start,
+      end,
+      currentOffset,
+      selectedBlockId,
+      blocks
+    );
+    this.showHyperlinkInput(existingLink);
+  }
+
+  getCommonHyperlinkInRange(
+    start: number,
+    end: number,
+    currentOffset: number,
+    selectedBlockId: string | null,
+    blocks: blockType
+  ): string | null {
+    // let offset = this.currentOffset;
+    let offset = currentOffset;
+    let index = 0;
+    if (selectedBlockId) {
+      index = blocks.findIndex(
+        (block: any) => block.dataId === selectedBlockId
+      );
     }
-  
-    // Method to save the current selection or caret position
-    private saveSelection() {
+    const pieces = blocks[index].pieces;
+    let commonLink: string | null = null;
+
+    for (let piece of pieces) {
+      const pieceEnd = offset + piece.text.length;
+      if (pieceEnd > start && offset < end) {
+        const pieceLink = piece.attributes.hyperlink || null;
+        if (commonLink === null) {
+          commonLink = pieceLink;
+        } else if (commonLink !== pieceLink) {
+          // Different hyperlinks in selection
+          return null;
+        }
+      }
+      offset = pieceEnd;
+    }
+    return commonLink;
+  }
+
+  showHyperlinkInput(existingLink: string | null): void {
+    // Get the elements
+    const hyperlinkContainer = document.getElementById("hyperlink-container");
+    const hyperlinkInput = document.getElementById(
+      "hyperlink-input"
+    ) as HTMLInputElement;
+    const applyButton = document.getElementById("apply-hyperlink");
+    const cancelButton = document.getElementById("cancel-hyperlink");
+
+    if (hyperlinkContainer && hyperlinkInput && applyButton && cancelButton) {
+      hyperlinkContainer.style.display = "block";
+
+      // position the container near the selection or toolbar
       const selection = window.getSelection();
       if (selection && selection.rangeCount > 0) {
-        this.savedRange = selection.getRangeAt(0).cloneRange();
+        const range = selection.getRangeAt(0);
+        const rect = range.getBoundingClientRect();
+        hyperlinkContainer.style.top = `${rect.bottom + window.scrollY + 5}px`;
+        hyperlinkContainer.style.left = `${rect.left + window.scrollX}px`;
       }
-    }
-  
-    public openHyperlinkModal() {
-      // If modal already exists, don't create it again
-      if (this.modal) return;
-  
-      // Create background overlay
-      this.overlay = document.createElement("div");
-      this.overlay.classList.add("modal-overlay");
-      document.body.appendChild(this.overlay);
-  
-      // Create modal structure
-      this.modal = document.createElement("div");
-      this.modal.classList.add("hyperlink-modal");
-  
-      const linkTextInput = document.createElement("input");
-      linkTextInput.type = "text";
-      linkTextInput.placeholder = "Enter link text";
-      linkTextInput.classList.add("hyperlink-input");
-  
-      const urlInput = document.createElement("input");
-      urlInput.type = "text";
-      urlInput.placeholder = "Enter URL";
-      urlInput.classList.add("hyperlink-input");
-  
-      const insertButton = document.createElement("button");
-      insertButton.textContent = "Insert/Update Link";
-      insertButton.onclick = () => this.insertHyperlink(linkTextInput.value, urlInput.value);
-  
-      const cancelButton = document.createElement("button");
-      cancelButton.textContent = "Close";
-      cancelButton.onclick = () => this.closeHyperlinkModal();
-  
-      // Append elements to modal
-      this.modal.appendChild(linkTextInput);
-      this.modal.appendChild(urlInput);
-      this.modal.appendChild(insertButton);
-      this.modal.appendChild(cancelButton);
-  
-      // Check if there is an anchor (hyperlink) in the saved selection
-      if (this.savedRange) {
-        const container = this.savedRange.commonAncestorContainer;
-        let anchor = container instanceof HTMLAnchorElement ? container : container.parentElement as HTMLAnchorElement;
-  
-        if (anchor && anchor.tagName === 'A') {
-          // If an anchor is found, populate the input fields for update functionality
-          this.currentAnchor = anchor;
-          linkTextInput.value = anchor.textContent || '';
-          urlInput.value = anchor.href;
+
+      // Set the existing link
+      hyperlinkInput.value = existingLink || "";
+
+      // Save the current selection
+      this.savedSelection = saveSelection(this.editorView.container);
+
+      // Show temporary selection
+      this.highlightSelection();
+
+      // Ensure the hyperlink input is focused
+      hyperlinkInput.focus();
+
+      // Remove any previous event listeners
+      applyButton.onclick = null;
+      cancelButton.onclick = null;
+
+      // Handle the 'Link' button
+      applyButton.onclick = () => {
+        const url = hyperlinkInput.value.trim();
+        if (url) {
+          this.applyHyperlink(url);
         }
-      }
-  
-      // Append the modal to the document body
-      document.body.appendChild(this.modal);
-  
-      // Style modal (add your own styling as needed)
-      this.modal.style.position = "fixed";
-      this.modal.style.top = "20%";
-      this.modal.style.left = "45%";
-      this.modal.style.backgroundColor = "#fff";
-      this.modal.style.zIndex = "1000"; // Ensure the modal is above other elements
-  
-      // Style overlay
-      this.overlay.style.position = "fixed";
-      this.overlay.style.top = "0";
-      this.overlay.style.left = "0";
-      this.overlay.style.width = "100%";
-      this.overlay.style.height = "100%";
-      this.overlay.style.backgroundColor = "rgba(0, 0, 0, 0.5)"; // Semi-transparent background
-      this.overlay.style.backdropFilter = "blur(5px)"; // Apply blur effect
-      this.overlay.style.zIndex = "999"; // Make sure overlay is below the modal
-    }
-  
-    private insertHyperlink(linkText: string, url: string) {
-      // Validate input
-      if (!linkText || !url) {
-        console.error("Both link text and URL are required.");
-        return;
-      }
-  
-      if (this.currentAnchor) {
-        // Update existing hyperlink
-        this.currentAnchor.href = url;
-        this.currentAnchor.textContent = linkText;
-        this.currentAnchor.target = "_blank"; // Ensure the link opens in a new tab
-        this.currentAnchor = null; // Clear reference after updating
-      } else if (this.savedRange) {
-        const selection = window.getSelection();
-  
-        // Restore the saved selection before inserting the hyperlink
-        selection?.removeAllRanges();
-        selection?.addRange(this.savedRange);
-  
-        // Create the anchor element
-        const anchor = document.createElement('a');
-        anchor.href = url;
-        anchor.textContent = linkText;
-        anchor.target = "_blank"; // Ensure the link opens in a new tab
-        anchor.style.textDecoration = 'none'; // Remove underline
-  
-        // Optional: Handle click event to prevent default navigation behavior if needed
-        anchor.onclick = (event: MouseEvent) => {
-          event.preventDefault();
-          window.open(anchor.href, '_blank'); // Open the URL in a new tab manually
-        };
-  
-        // Use DocumentFragment for safer insertion
-        const fragment = document.createDocumentFragment();
-        fragment.appendChild(anchor);
-  
-        try {
-          // Insert the fragment into the saved range
-          this.savedRange.deleteContents();
-          this.savedRange.insertNode(fragment);
-  
-          // Move the cursor right after the inserted link
-          this.savedRange.setStartAfter(anchor);
-          this.savedRange.setEndAfter(anchor);
-  
-          // Clear the saved range
-          this.savedRange = null;
-        } catch (error) {
-          console.error("Error inserting hyperlink:", error);
-        }
-      } else {
-        console.error("No saved selection range found.");
-      }
-  
-      this.closeHyperlinkModal();
-    }
-  
-    private closeHyperlinkModal() {
-      if (this.modal) {
-        document.body.removeChild(this.modal);
-        this.modal = null;
-      }
-      if (this.overlay) {
-        document.body.removeChild(this.overlay);
-        this.overlay = null;
-      }
+        hyperlinkContainer.style.display = "none";
+      };
+
+      // Handle the 'Unlink' button
+      cancelButton.onclick = () => {
+        // console.log("yyy", { remove: this.document.dataIds });
+        this.removeHyperlink();
+        hyperlinkContainer.style.display = "none";
+      };
     }
   }
+
+  highlightSelection(): void {
+    // Remove any existing temporary highlights
+    this.removeHighlightSelection();
+
+    const selection = window.getSelection();
+    if (selection && selection.rangeCount > 0) {
+      const range = selection.getRangeAt(0);
+
+      // Create a wrapper span
+      const span = document.createElement("span");
+      span.className = "temporary-selection-highlight";
+
+      // Extract the selected content and wrap it
+      span.appendChild(range.extractContents());
+      range.insertNode(span);
+
+      // Adjust the selection to encompass the new span
+      selection.removeAllRanges();
+      const newRange = document.createRange();
+      newRange.selectNodeContents(span);
+      selection.addRange(newRange);
+    }
+  }
+
+  removeHighlightSelection(): void {
+    const highlights = this.editorContainer?.querySelectorAll(
+      "span.temporary-selection-highlight"
+    );
+    highlights?.forEach((span) => {
+      const parent = span.parentNode;
+      if (parent) {
+        while (span.firstChild) {
+          parent.insertBefore(span.firstChild, span);
+        }
+        parent.removeChild(span);
+      }
+    });
+  }
+
+
+  applyHyperlink(url: string): void {
+    // Remove any existing temporary highlights
+    this.removeHighlightSelection();
+
+    // Restore the selection
+    restoreSelection(this.editorView.container, this.savedSelection);
+
+    const [start, end] = getSelectionRange(this.editorView);
+    if (start < end) {
+        console.log('yyy',this.document.blocks)
+        console.log('yyy',this.document.selectedBlockId)
+        console.log('yyy',this.document.dataIds)
+        
+        if (this.document.dataIds.length > 1) {
+            this.document.blocks.forEach((block: any) => {
+                if (this.document.dataIds.includes(block.dataId)) {
+                    this.document.selectedBlockId = block.dataId;
+                    let countE = 0;
+                    block.pieces.forEach((obj: any) => {
+                        countE += obj.text.length;
+                    })
+                    let countS = start - countE;
+                    // this.document.applyHyperlinkRange(countS, countE,url);
+                    this.document.formatAttribute(countS, countE, 'hyperlink', url);
+                }
+            })
+        } else {
+            // this.document.applyHyperlinkRange(start, end,url);
+            this.document.formatAttribute(start, end, 'hyperlink', url);
+        }
+      
+
+
+        // this.document.applyHyperlinkRange(start, end, url);
+
+        this.editorView.render();
+        // Restore selection and focus
+        restoreSelection(this.editorView.container, this.savedSelection);
+        this.editorView.container.focus();
+    }
+    this.savedSelection = null;
+}
+
+removeHyperlink(): void {
+    // Remove any existing temporary highlights
+    this.removeHighlightSelection();
+
+    // Restore the selection
+    restoreSelection(this.editorView.container, this.savedSelection);
+
+    const [start, end] = getSelectionRange(this.editorView);
+    console.log('yyy',{remove:this.document.dataIds})
+    if (start < end) {
+        // this.document.removeHyperlinkRange(start, end);
+
+
+
+
+
+
+        if (this.document.dataIds.length > 1) {
+            this.document.blocks.forEach((block: any) => {
+                if (this.document.dataIds.includes(block.dataId)) {
+                    this.document.selectedBlockId = block.dataId;
+                    let countE = 0;
+                    block.pieces.forEach((obj: any) => {
+                        countE += obj.text.length;
+                    })
+                    let countS = start - countE;
+                    // this.document.removeHyperlinkRange(countS, countE);
+                    this.document.formatAttribute(countS, countE, 'hyperlink', false);
+
+                }
+            })
+        } else {
+            this.document.formatAttribute(start, end, 'hyperlink', false);
+
+            // this.document.removeHyperlinkRange(start, end);
+        }
+      
+
+
+
+        this.editorView.render();
+        // Restore selection and focus
+        restoreSelection(this.editorView.container, this.savedSelection);
+        this.editorView.container.focus();
+
+
+
+
+    }
+    this.savedSelection = null;
+}
+
+
+showHyperlinkViewButton(link:string | "") : void{
+    const viewHyperlinkContainer = document.getElementById('hyperlink-container-view') as HTMLDivElement;
+    const hyperLinkAnchor = document.getElementById('hyperlink-view-link') as HTMLAnchorElement;
+  
+
+    if (viewHyperlinkContainer && hyperLinkAnchor) {
+        viewHyperlinkContainer.style.display = 'block';
+
+        // position the container near the selection or toolbar
+        const selection = window.getSelection();
+        if (selection) {
+            const range = selection.getRangeAt(0);
+            const rect = range.getBoundingClientRect();
+            viewHyperlinkContainer.style.top = `${rect.bottom + window.scrollY + 5}px`;
+            viewHyperlinkContainer.style.left = `${rect.left + window.scrollX}px`;
+        }
+
+        // Set the existing link
+        if(link){
+            hyperLinkAnchor.innerText = link || '';
+            hyperLinkAnchor.href = link || '';
+        }
+    }
+}
+
+ hideHyperlinkViewButton(){
+    const hyperlinkContainer = document.getElementById('hyperlink-container-view');
+    if(hyperlinkContainer){
+        hyperlinkContainer.style.display = 'none';
+    }
+}
+
+
+}
+
+export default HyperlinkHandler;
