+++ src/index.ts
@@ -1,10 +1,10 @@
 import TextDocument from "./textDocument";
 import EditorView from "./view/editorView";
 import ToolbarView from "./view/toolbarView";
+import HyperlinkHandler  from "./handlers/hyperlink";
 import Piece from "./piece";
 import { saveSelection, restoreSelection } from "./utils/selectionManager";
 import { parseHtmlToPieces } from "./utils/parseHtml";
-import { showHyperlinkViewButton, hideHyperlinkViewButton } from './attributes/hyperLink'
 import { createEditor } from "./config/editorConfig";
 import "./styles/text-igniter.css"
 
@@ -17,6 +17,7 @@ class TextIgniter {
     document: TextDocument;
     editorView: EditorView;
     toolbarView: ToolbarView;
+    hyperlinkHandler:HyperlinkHandler;
     currentAttributes: CurrentAttributeDTO;
     manualOverride: boolean;
     lastPiece: Piece | null;
@@ -38,6 +39,7 @@ class TextIgniter {
         this.document = new TextDocument();
         this.editorView = new EditorView(this.editorContainer, this.document);
         this.toolbarView = new ToolbarView(this.toolbarContainer);
+        this.hyperlinkHandler = new HyperlinkHandler(this.editorContainer,this.editorView,this.document);
         this.currentAttributes = { bold: false, italic: false, underline: false, undo: false, redo: false, hyperlink: false };
         this.manualOverride = false;
         this.lastPiece = null;
@@ -110,7 +112,7 @@ class TextIgniter {
             // this.document.setAlignment('right', this.document.selectedBlockId);
         });
 
-        document.addEventListener('keydown', (e) => {
+        this.editorContainer.addEventListener('keydown', (e) => {
             if ((e.ctrlKey || e.metaKey) && !e.altKey) {
                 const key = e.key.toLowerCase();
                 if (['b', 'i', 'u', 'h'].includes(key)) {
@@ -320,8 +322,7 @@ class TextIgniter {
                     this.document.redo();
                     break;
                 case 'hyperlink':
-                        const existingLink = this.document.getCommonHyperlinkInRange(start, end);
-                        this.showHyperlinkInput(existingLink);
+                        this.hyperlinkHandler.hanldeHyperlinkClick(start,end,this.document.currentOffset,this.document.selectedBlockId,this.document.blocks);
                     break;
             }
         } else {
@@ -332,184 +333,8 @@ class TextIgniter {
         this.toolbarView.updateActiveStates(this.currentAttributes);
     }
 
-    showHyperlinkInput(existingLink: string | null): void {
-        // Get the elements
-        const hyperlinkContainer = document.getElementById('hyperlink-container');
-        const hyperlinkInput = document.getElementById('hyperlink-input') as HTMLInputElement;
-        const applyButton = document.getElementById('apply-hyperlink');
-        const cancelButton = document.getElementById('cancel-hyperlink');
 
-        if (hyperlinkContainer && hyperlinkInput && applyButton && cancelButton) {
-            hyperlinkContainer.style.display = 'block';
-
-            // position the container near the selection or toolbar
-            const selection = window.getSelection();
-            if (selection && selection.rangeCount > 0) {
-                const range = selection.getRangeAt(0);
-                const rect = range.getBoundingClientRect();
-                hyperlinkContainer.style.top = `${rect.bottom + window.scrollY + 5}px`;
-                hyperlinkContainer.style.left = `${rect.left + window.scrollX}px`;
-            }
-
-            // Set the existing link
-            hyperlinkInput.value = existingLink || '';
-
-            // Save the current selection
-            this.savedSelection = saveSelection(this.editorView.container);
-
-            // Show temporary selection 
-            this.highlightSelection();
-
-            // Ensure the hyperlink input is focused
-            hyperlinkInput.focus();
-
-            // Remove any previous event listeners
-            applyButton.onclick = null;
-            cancelButton.onclick = null;
-
-            // Handle the 'Link' button
-            applyButton.onclick = () => {
-                const url = hyperlinkInput.value.trim();
-                if (url) {
-                    this.applyHyperlink(url);
-                }
-                hyperlinkContainer.style.display = 'none';
-            };
-
-            // Handle the 'Unlink' button
-            cancelButton.onclick = () => {
-        console.log('zzz',{remove:this.document.dataIds})
-                this.removeHyperlink();
-                hyperlinkContainer.style.display = 'none';
-            };
-        }
-    }
-
-    highlightSelection(): void {
-        // Remove any existing temporary highlights
-        this.removeHighlightSelection();
-
-        const selection = window.getSelection();
-        if (selection && selection.rangeCount > 0) {
-            const range = selection.getRangeAt(0);
-
-            // Create a wrapper span
-            const span = document.createElement('span');
-            span.className = 'temporary-selection-highlight';
-
-            // Extract the selected content and wrap it
-            span.appendChild(range.extractContents());
-            range.insertNode(span);
-
-            // Adjust the selection to encompass the new span
-            selection.removeAllRanges();
-            const newRange = document.createRange();
-            newRange.selectNodeContents(span);
-            selection.addRange(newRange);
-        }
-    }
-
-    removeHighlightSelection(): void {
-        const highlights = this.editorContainer?.querySelectorAll('span.temporary-selection-highlight');
-        highlights?.forEach((span) => {
-            const parent = span.parentNode;
-            if (parent) {
-                while (span.firstChild) {
-                    parent.insertBefore(span.firstChild, span);
-                }
-                parent.removeChild(span);
-            }
-        });
-    }
-
-    applyHyperlink(url: string): void {
-        // Remove any existing temporary highlights
-        this.removeHighlightSelection();
-
-        // Restore the selection
-        restoreSelection(this.editorView.container, this.savedSelection);
-
-        const [start, end] = this.getSelectionRange();
-        if (start < end) {
-            console.log('zzz',this.document.blocks)
-            console.log('zzz',this.document.selectedBlockId)
-            console.log('zzz',this.document.dataIds)
-            
-            if (this.document.dataIds.length > 1) {
-                this.document.blocks.forEach((block: any) => {
-                    if (this.document.dataIds.includes(block.dataId)) {
-                        this.document.selectedBlockId = block.dataId;
-                        let countE = 0;
-                        block.pieces.forEach((obj: any) => {
-                            countE += obj.text.length;
-                        })
-                        let countS = start - countE;
-                        this.document.applyHyperlinkRange(countS, countE,url);
-                    }
-                })
-            } else {
-                this.document.applyHyperlinkRange(start, end,url);
-            }
-          
-
-
-            // this.document.applyHyperlinkRange(start, end, url);
-
-            this.editorView.render();
-            // Restore selection and focus
-            restoreSelection(this.editorView.container, this.savedSelection);
-            this.editorView.container.focus();
-        }
-        this.savedSelection = null;
-    }
-
-    removeHyperlink(): void {
-        // Remove any existing temporary highlights
-        this.removeHighlightSelection();
-
-        // Restore the selection
-        restoreSelection(this.editorView.container, this.savedSelection);
-
-        const [start, end] = this.getSelectionRange();
-        console.log('zzz',{remove:this.document.dataIds})
-        if (start < end) {
-            // this.document.removeHyperlinkRange(start, end);
-
-
-
-
-
-
-            if (this.document.dataIds.length > 1) {
-                this.document.blocks.forEach((block: any) => {
-                    if (this.document.dataIds.includes(block.dataId)) {
-                        this.document.selectedBlockId = block.dataId;
-                        let countE = 0;
-                        block.pieces.forEach((obj: any) => {
-                            countE += obj.text.length;
-                        })
-                        let countS = start - countE;
-                        this.document.removeHyperlinkRange(countS, countE);
-                    }
-                })
-            } else {
-                this.document.removeHyperlinkRange(start, end);
-            }
-          
-
-
-
-            this.editorView.render();
-            // Restore selection and focus
-            restoreSelection(this.editorView.container, this.savedSelection);
-            this.editorView.container.focus();
-
-
-
-
-        }
-        this.savedSelection = null;
-    }
+       
 
     handleSelectionChange(): void {
         this.syncCurrentAttributesWithCursor();
@@ -775,10 +600,10 @@ class TextIgniter {
                 // Show below link..
                 const hyperlink = piece?.attributes.hyperlink;
                 if (hyperlink && typeof hyperlink === 'string') {
-                    showHyperlinkViewButton(hyperlink);
+                    this.hyperlinkHandler.showHyperlinkViewButton(hyperlink);
                 }
                 else {
-                    hideHyperlinkViewButton()
+                    this.hyperlinkHandler.hideHyperlinkViewButton()
                 }
             } else {
                 if (!this.manualOverride) {
