+++ src/types/pieces.ts
@@ -1,0 +1,1 @@
+export type blockType = any;
